+++ onoro/src/lib.rs
@@ -10,7 +10,6 @@ mod r#move;
 mod onoro;
 mod onoro_defs;
 mod onoro_impl;
-mod onoro_initialize;
 mod onoro_state;
 mod onoro_util;
 mod onoro_view;
@@ -24,7 +23,6 @@ pub use color_print::*;
 pub use compress::*;
 pub use onoro_defs::*;
 pub use onoro_impl::*;
-pub use onoro_initialize::*;
 pub use onoro_view::*;
 pub use packed_idx::*;
 pub use r#move::*;
