+++ onoro/src/onoro_impl.rs
@@ -13,7 +13,7 @@ use crate::{
   groups::{C2, D3, D6, K4},
   make_onoro_error,
   util::broadcast_u8_to_u64,
-  Color, Colored, Onoro, OnoroExt, OnoroPawn, PawnColor, TileState,
+  Color, Colored, Onoro, OnoroPawn, PawnColor, TileState,
 };
 
 use super::{
@@ -605,6 +605,18 @@ impl<const N: usize, const N2: usize, const ADJ_CNT_SIZE: usize> Onoro
   type Move = Move;
   type Pawn = Pawn;
 
+  unsafe fn new() -> Self {
+    Self {
+      pawn_poses: [PackedIdx::null(); N],
+      state: OnoroState::new(),
+      sum_of_mass: HexPos::zero().into(),
+    }
+  }
+
+  fn pawns_per_player() -> usize {
+    N / 2
+  }
+
   fn turn(&self) -> PawnColor {
     if self.onoro_state().black_turn() {
       PawnColor::Black
@@ -678,22 +690,6 @@ impl<const N: usize, const N2: usize, const ADJ_CNT_SIZE: usize> Onoro
         self.move_pawn(from_idx as usize, to);
       }
     }
-  }
-}
-
-impl<const N: usize, const N2: usize, const ADJ_CNT_SIZE: usize> OnoroExt
-  for OnoroImpl<N, N2, ADJ_CNT_SIZE>
-{
-  unsafe fn new() -> Self {
-    Self {
-      pawn_poses: [PackedIdx::null(); N],
-      state: OnoroState::new(),
-      sum_of_mass: HexPos::zero().into(),
-    }
-  }
-
-  fn pawns_per_player() -> usize {
-    N / 2
   }
 }
 
@@ -1106,7 +1102,7 @@ impl<const N: usize, const N2: usize, const ADJ_CNT_SIZE: usize> GameMoveGenerat
 
 #[cfg(test)]
 mod tests {
-  use crate::{onoro_defs::Onoro8, packed_idx::PackedIdx, Onoro, OnoroExt};
+  use crate::{onoro_defs::Onoro8, packed_idx::PackedIdx, Onoro};
 
   #[test]
   fn test_get_tile() {
